Removed warning note about no flash messages being displayed, since as of version 0.8 they are displayed using a partial.
The typeOfScaffold had two different options set as the default selected.  Now only "Model, Views and Controller" is selected by default.

Moved the example text out of the CFIF block as it should be displayed each time as a guide for the user to create a new scaffold, not just the first time they run the process.